adding mockchain.hs in pretty
