+++ src/Cooked/Pretty.hs
@@ -48,6 +48,7 @@ module Cooked.Pretty (module X) where
 
 import Cooked.Pretty.Common as X
 import Cooked.Pretty.Hashable as X
+import Cooked.Pretty.MockChain as X ()
 import Cooked.Pretty.Options as X
 import Cooked.Pretty.Plutus as X ()
 import Cooked.Pretty.Skeleton as X ()
