+++ src/Cooked/Pretty/Skeleton.hs
@@ -1,23 +1,10 @@
 {-# LANGUAGE DeriveFunctor #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
--- | This module implements pretty-printing for Cooked structures such as
--- skeletons and chain state.
---
--- It contains orphaned instances of 'PrettyCooked' for Cooked datatypes. They
--- cannot be provided in "Cooked.Pretty.Class" because of dependency cycles and,
--- for ease of maintainability, we chose to centralize all pretty-printing
--- related code in submodules of "Cooked.Pretty" instead of having
--- 'PrettyCooked' instances scattered around.
---
--- Some structure require additional arguments to be pretty-printed and have
--- therefore no instances 'PrettyCooked' (for example 'TxSkel' needs some
--- 'SkelContext').
-module Cooked.Pretty.Skeleton () where
+-- | This module implements 'PrettyCooked', 'PrettyCookedL' and 'PrettyCookedM'
+-- instances for 'TxSkel's and its components.
+module Cooked.Pretty.Skeleton (Contextualized (..), CollateralInput (..)) where
 
-import Cooked.MockChain.BlockChain
-import Cooked.MockChain.Direct
-import Cooked.MockChain.UtxoState
 import Cooked.Output
 import Cooked.Pretty.Common
 import Cooked.Pretty.Options
@@ -25,11 +12,9 @@ import Cooked.Pretty.Plutus ()
 import Cooked.Skeleton
 import Cooked.Wallet
 import Data.Default
-import Data.Function (on)
-import Data.List qualified as List
 import Data.Map (Map)
 import Data.Map qualified as Map
-import Data.Maybe (catMaybes, fromMaybe, mapMaybe)
+import Data.Maybe (catMaybes, mapMaybe)
 import Data.Set qualified as Set
 import Optics.Core
 import Plutus.Script.Utils.Data qualified as Script
@@ -50,117 +35,6 @@ instance PrettyCooked TxSkelOutDatum where
 
 instance PrettyCooked Wallet where
   prettyCookedOpt opts = prettyHash opts . walletPKHash
-
-instance PrettyCooked MockChainError where
-  prettyCookedOpt opts (MCEValidationError plutusPhase plutusError) =
-    PP.vsep ["Validation error " <+> prettyCookedOpt opts plutusPhase, PP.indent 2 (prettyCookedOpt opts plutusError)]
-  -- Here we don't print the skel because we lack its context and this error is
-  -- printed alongside the skeleton when a test fails
-  prettyCookedOpt opts (MCEUnbalanceable balWallet missingValue _) =
-    prettyItemize
-      opts
-      "Unbalanceable:"
-      "-"
-      [ prettyCookedOpt opts balWallet <+> "does not have enough funds",
-        if missingValue == mempty
-          then "Not enough funds to sustain the minimal ada of the return utxo"
-          else "Unable to find" <+> prettyCookedOpt opts missingValue
-      ]
-  prettyCookedOpt opts (MCENoSuitableCollateral fee percentage colVal) =
-    prettyItemize
-      opts
-      "No suitable collateral"
-      "-"
-      [ "Fee was" <+> prettyCookedOpt opts fee,
-        "Percentage in params was" <+> prettyCookedOpt opts percentage,
-        "Resulting minimal collateral value was" <+> prettyCookedOpt opts colVal
-      ]
-  prettyCookedOpt opts (MCEGenerationError (ToCardanoError msg cardanoError)) =
-    prettyItemize
-      opts
-      "Transaction generation error:"
-      "-"
-      [PP.pretty msg :: DocCooked, PP.pretty cardanoError]
-  prettyCookedOpt opts (MCEGenerationError (GenerateTxErrorGeneral msgs)) =
-    prettyItemize opts "Transaction generation error:" "-" [PP.pretty msgs :: DocCooked]
-  prettyCookedOpt opts (MCEGenerationError (TxBodyError msg err)) =
-    prettyItemize opts "Transaction generation error:" "-" [PP.pretty msg :: DocCooked, PP.viaShow err]
-  prettyCookedOpt opts (MCEUnknownOutRefError msg txOutRef) =
-    prettyItemize opts "Unknown transaction output ref:" "-" [PP.pretty msg, prettyCookedOpt opts txOutRef]
-  prettyCookedOpt _ (FailWith msg) = "Failed with:" <+> PP.pretty msg
-  prettyCookedOpt opts (MCEUnknownValidator msg valHash) =
-    prettyItemize
-      opts
-      "Unknown validator hash:"
-      "-"
-      [PP.pretty msg, "hash:" <+> prettyHash opts valHash]
-  prettyCookedOpt opts (MCEUnknownDatum msg dHash) =
-    prettyItemize
-      opts
-      "Unknown datum hash:"
-      "-"
-      [PP.pretty msg, "hash:" <+> prettyHash opts dHash]
-
-instance (Show a) => PrettyCooked (a, UtxoState) where
-  prettyCookedOpt opts (res, state) =
-    prettyItemize
-      opts
-      "End state:"
-      "-"
-      ["Returns:" <+> PP.viaShow res, prettyCookedOpt opts state]
-
-instance (Show a) => PrettyCooked (MockChainReturn a UtxoState) where
-  prettyCookedOpt opts' (res, MockChainBook entries ((`addHashNames` opts') -> opts)) =
-    let mcEndResult = case res of
-          Left err -> "🔴" <+> prettyCookedOpt opts err
-          Right (a, s) -> "🟢" <+> prettyCookedOpt opts (a, s)
-     in PP.vsep $ if pcOptPrintLog opts then [prettyCookedOpt opts entries, mcEndResult] else [mcEndResult]
-
-instance PrettyCooked [MockChainLogEntry] where
-  prettyCookedOpt opts = ("📘" <+>) . prettyItemize opts "MockChain run log:" "⁍"
-
--- | This prints a 'MockChainLogEntry'. In the log, we know a transaction has
--- been validated if the 'MCLogSubmittedTxSkel' is followed by a 'MCLogNewTx'.
-instance PrettyCooked MockChainLogEntry where
-  prettyCookedOpt opts (MCLogAdjustedTxSkelOut skelOut newAda) =
-    "The ADA amount of "
-      <> prettyCookedOpt opts skelOut
-      <> " has been automatically adjusted to "
-      <> prettyCookedOpt opts (Script.toValue newAda)
-  prettyCookedOpt opts (MCLogSubmittedTxSkel outputs datums skel) = "Submitted:" <+> prettyCookedOpt opts (Contextualized outputs datums skel)
-  prettyCookedOpt opts (MCLogAdjustedTxSkel outputs datums skel fee mCollaterals) =
-    let mCollateralsDoc =
-          ( \(collaterals, returnWallet) ->
-              [ prettyItemize opts "Collateral inputs:" "-" (Contextualized outputs datums . CollateralInput <$> Set.toList collaterals),
-                "Return collateral target:" <+> prettyCookedOpt opts returnWallet
-              ]
-          )
-            <$> mCollaterals
-     in prettyItemize
-          opts
-          "Adjusted:"
-          "-"
-          $ [ prettyCookedOpt opts (Contextualized outputs datums skel),
-              "Fee:" <+> prettyCookedOpt opts (Script.lovelace fee)
-            ]
-            ++ fromMaybe [] mCollateralsDoc
-  prettyCookedOpt opts (MCLogNewTx txId) = "New transaction:" <+> prettyHash opts txId
-  prettyCookedOpt opts (MCLogDiscardedUtxos n s) = prettyCookedOpt opts n <+> "balancing utxos were discarded:" <+> PP.pretty s
-  prettyCookedOpt opts (MCLogUnusedCollaterals (Left cWallet)) =
-    "Specific request to fetch collateral utxos from "
-      <> prettyCookedOpt opts cWallet
-      <> " has been disregarded because the transaction does not require collaterals"
-  prettyCookedOpt opts (MCLogUnusedCollaterals (Right (length -> n))) =
-    "Specific request to fetch collateral utxos from the given set of "
-      <> prettyCookedOpt opts n
-      <> " elements has been disregarded because the transaction does not require collaterals"
-  prettyCookedOpt opts (MCLogAddedReferenceScript red oRef sHash) =
-    "A reference script located in "
-      <> prettyCookedOpt opts oRef
-      <> " has been automatically associated to redeemer "
-      <> prettyItemizeNoTitle opts "-" red
-      <> " for script "
-      <> prettyHash opts sHash
 
 data Contextualized a = Contextualized
   { _ctxOutputs :: Map Api.TxOutRef Api.TxOut,
@@ -392,22 +266,6 @@ instance PrettyCookedM TxSkelOutDatum where
         <> "):"
         <+> PP.align (prettyCookedOpt opts txSkelOutDatum)
 
--- | Resolves a "TxOutRef" from a given context, builds a doc cooked for its
--- address and value, and also builds a possibly empty list for its datum and
--- reference script when they exist.
-utxoToPartsAsDocCooked :: PrettyCookedOpts -> Contextualized Api.TxOutRef -> Maybe (DocCooked, DocCooked, [DocCooked])
-utxoToPartsAsDocCooked opts ctx =
-  ( \(output, txSkelOutDatum) ->
-      ( prettyCookedOpt opts (outputAddress output),
-        prettyCookedOpt opts (outputValue output),
-        catMaybes
-          [ prettyCookedOptM opts txSkelOutDatum,
-            getReferenceScriptDoc opts output
-          ]
-      )
-  )
-    <$> lookupOutput ctx
-
 newtype CollateralInput = CollateralInput {unCollateralInput :: Api.TxOutRef}
 
 -- | Prints a collateral input within a certain 'SkelContext'
@@ -449,16 +307,30 @@ instance PrettyCookedM (Contextualized ReferenceInput) where
 getReferenceScriptDoc :: (IsAbstractOutput output, Script.ToScriptHash (ReferenceScriptType output)) => PrettyCookedOpts -> output -> Maybe DocCooked
 getReferenceScriptDoc opts output = ("Reference script hash:" <+>) . prettyHash opts . Script.toScriptHash <$> output ^. outputReferenceScriptL
 
+-- | Resolves a "TxOutRef" from a given context, builds a doc cooked for its
+-- address and value, and also builds a possibly empty list for its datum and
+-- reference script when they exist.
+utxoToPartsAsDocCooked :: PrettyCookedOpts -> Contextualized Api.TxOutRef -> Maybe (DocCooked, DocCooked, [DocCooked])
+utxoToPartsAsDocCooked opts ctx =
+  ( \(output, txSkelOutDatum) ->
+      ( prettyCookedOpt opts (outputAddress output),
+        prettyCookedOpt opts (outputValue output),
+        catMaybes
+          [ prettyCookedOptM opts txSkelOutDatum,
+            getReferenceScriptDoc opts output
+          ]
+      )
+  )
+    <$> lookupOutput ctx
+
 lookupOutput :: Contextualized Api.TxOutRef -> Maybe (Api.TxOut, TxSkelOutDatum)
 lookupOutput (Contextualized managedTxOuts managedTxSkelOutDatums txOutRef) = do
   output <- Map.lookup txOutRef managedTxOuts
-  return
-    ( output,
-      case outputOutputDatum output of
-        Api.OutputDatum datum -> Map.findWithDefault TxSkelOutNoDatum (Script.datumHash datum) managedTxSkelOutDatums
-        Api.OutputDatumHash datumHash -> Map.findWithDefault TxSkelOutNoDatum datumHash managedTxSkelOutDatums
-        Api.NoOutputDatum -> TxSkelOutNoDatum
-    )
+  (output,)
+    <$> case outputOutputDatum output of
+      Api.OutputDatum datum -> Map.lookup (Script.datumHash datum) managedTxSkelOutDatums
+      Api.OutputDatumHash datumHash -> Map.lookup datumHash managedTxSkelOutDatums
+      Api.NoOutputDatum -> return TxSkelOutNoDatum
 
 -- | Pretty-print a list of transaction skeleton options, only printing an
 -- option if its value is non-default.
@@ -542,94 +414,3 @@ instance PrettyCookedL TxOpts where
         prettyAnchorResolution :: AnchorResolution -> DocCooked
         prettyAnchorResolution AnchorResolutionHttp = "Resolve anchor url with an (unsafe) http connection"
         prettyAnchorResolution (AnchorResolutionLocal urlMap) = prettyItemize @[DocCooked] opts "Resolve anchor url with the following table keys" "-" (PP.viaShow <$> Map.keys urlMap)
-
--- | Pretty print a 'UtxoState'. Print the known wallets first, then unknown
--- pubkeys, then scripts.
-instance PrettyCooked UtxoState where
-  prettyCookedOpt opts =
-    prettyItemize opts "UTxO state:" "•"
-      . map (uncurry (prettyAddressState opts))
-      . List.sortBy addressOrdering
-      . Map.toList
-      . utxoState
-    where
-      addressOrdering :: (Api.Address, a) -> (Api.Address, a) -> Ordering
-      addressOrdering
-        (a1@(Api.Address (Api.PubKeyCredential pkh1) _), _)
-        (a2@(Api.Address (Api.PubKeyCredential pkh2) _), _) =
-          case (walletPKHashToId pkh1, walletPKHashToId pkh2) of
-            (Just i, Just j) -> compare i j
-            (Just _, Nothing) -> LT
-            (Nothing, Just _) -> GT
-            (Nothing, Nothing) -> compare a1 a2
-      addressOrdering
-        (Api.Address (Api.PubKeyCredential _) _, _)
-        (Api.Address (Api.ScriptCredential _) _, _) = LT
-      addressOrdering (a1, _) (a2, _) = compare a1 a2
-
--- | Pretty prints the state of an address, that is the list of UTxOs (including
--- value and datum), grouped
-prettyAddressState :: PrettyCookedOpts -> Api.Address -> UtxoPayloadSet -> DocCooked
-prettyAddressState opts address payloadSet =
-  prettyItemize
-    opts
-    (prettyCookedOpt opts address)
-    "-"
-    ( mapMaybe (prettyPayloadGrouped opts)
-        . group
-        . List.sortBy (compare `on` (Api.lovelaceValueOf . utxoPayloadValue))
-        . utxoPayloadSet
-        $ payloadSet
-    )
-  where
-    similar :: UtxoPayload -> UtxoPayload -> Bool
-    similar
-      (UtxoPayload _ value1 skelOutDatum1 refScript1)
-      (UtxoPayload _ value2 skelOutDatum2 refScript2) =
-        value1 == value2
-          && skelOutDatum1 == skelOutDatum2
-          && refScript1 == refScript2
-    group :: [UtxoPayload] -> [[UtxoPayload]]
-    group =
-      case pcOptPrintTxOutRefs opts of
-        PCOptTxOutRefsFull -> map (: [])
-        _ -> List.groupBy similar
-
--- | Pretty prints payloads (datum and value corresponding to 1 UTxO) grouped
--- together when they carry same value and datum
-prettyPayloadGrouped :: PrettyCookedOpts -> [UtxoPayload] -> Maybe DocCooked
-prettyPayloadGrouped _ [] = Nothing
-prettyPayloadGrouped opts [payload] =
-  prettyPayload
-    opts
-    (pcOptPrintTxOutRefs opts /= PCOptTxOutRefsHidden)
-    payload
-prettyPayloadGrouped opts (payload : rest) =
-  let cardinality = 1 + length rest
-   in (PP.parens ("×" <> prettyCookedOpt opts cardinality) <+>)
-        <$> prettyPayload opts False payload
-
--- | Optionally prints a 'UtxoPayload' with an option piloting whether
--- 'Api.TxOutRef's should be shown.
-prettyPayload :: PrettyCookedOpts -> Bool -> UtxoPayload -> Maybe DocCooked
-prettyPayload
-  opts
-  showTxOutRef
-  ( UtxoPayload
-      { utxoPayloadTxOutRef,
-        utxoPayloadValue,
-        utxoPayloadSkelOutDatum,
-        utxoPayloadReferenceScript
-      }
-    ) =
-    case catMaybes
-      [ if showTxOutRef
-          then Just $ prettyCookedOpt opts utxoPayloadTxOutRef
-          else Nothing,
-        Just (prettyCookedOpt opts utxoPayloadValue),
-        prettyCookedOptM opts utxoPayloadSkelOutDatum,
-        ("Reference script hash:" <+>) . prettyHash opts <$> utxoPayloadReferenceScript
-      ] of
-      [] -> Nothing
-      [doc] -> Just $ PP.align doc
-      docs -> Just . PP.align . PP.vsep $ docs
